Consolidate merge script into plotting CLI and update docs
